payment screen changed
